+++ python/ray/llm/tests/serve/cpu/deployments/llm/test_llm_server.py
@@ -20,7 +20,7 @@ from ray.llm._internal.serve.deployments.llm.llm_server import (
 from ray.llm._internal.serve.configs.constants import MODEL_RESPONSE_BATCH_TIMEOUT_MS
 from ray.llm.tests.serve.mocks.mock_vllm_engine import MockVLLMEngine
 
-# Add this to your test file
+
 async def create_server(*args, **kwargs):
     server = LLMServer.__new__(LLMServer)
     await server.__init__(*args, **kwargs)
